requirements.txt
alembic==1.16.5
anthropic>=0.30.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
backports.asyncio.runner==1.2.0
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.1.8
cryptography==46.0.7
ecdsa==0.19.2
exceptiongroup==1.3.1
Faker==37.12.0
fastapi==0.128.8
google-api-core==2.30.3
google-auth==2.49.2
google-cloud-vision==3.13.0
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.80.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.1.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy==2.0.2
packaging==26.0
pandas==2.3.3
passlib==1.7.4
pdf2image==1.17.0
pillow==11.3.0
pillow_heif==1.1.1
pluggy==1.6.0
probableparsing==0.0.1
proto-plus==1.27.2
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
Pygments==2.20.0
pytesseract==0.3.13
pytest==8.4.2
pytest-asyncio==1.2.0
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2026.1.post1
PyYAML==6.0.3
RapidFuzz==3.13.0
reportlab==4.4.10
requests==2.32.5
rich==14.3.3
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.49
starlette==0.49.3
tomli==2.4.1
typer==0.23.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
usaddress==0.5.16
uvicorn==0.39.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
